+++ src/Entity/CrowdfundingMedia.php
@@ -11,108 +11,31 @@
 namespace c975L\ShopBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
-use c975L\ShopBundle\Repositiry\CrowdfundingMediaRepository;
+use c975L\ShopBundle\Entity\Media;
+use c975L\ShopBundle\Entity\Crowdfunding;
 use Symfony\Component\HttpFoundation\File\File;
 use Vich\UploaderBundle\Mapping\Annotation as Vich;
 
-#[ORM\Entity(repositoryClass: CrowdfundingMediaRepository::class)]
-#[ORM\Table(name: 'shop_crowdfunding_media')]
+#[ORM\Entity]
 #[Vich\Uploadable]
-class CrowdfundingMedia
+class CrowdfundingMedia extends Media
 {
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column]
-    private ?int $id = null;
-
-    #[ORM\Column(nullable: true)]
-    private ?int $position = null;
-
-    #[ORM\Column(length: 255, nullable: true)]
-    private ?string $name = null;
-
-    #[ORM\Column(nullable: true)]
-    private ?int $size = null;
-
     #[Vich\UploadableField(mapping: 'crowdfundings', fileNameProperty: 'name', size: 'size')]
-    private ?File $file = null;
+    protected ?File $file = null;
 
-    #[ORM\Column]
-    private ?\DateTimeImmutable $updatedAt = null;
-
-    #[ORM\ManyToOne(inversedBy: 'medias')]
+    #[ORM\ManyToOne(targetEntity: Crowdfunding::class, inversedBy: 'medias')]
+    #[ORM\JoinColumn(nullable: true)]
     private ?Crowdfunding $crowdfunding = null;
 
-    public function getId(): ?int
+    public function getMappingName(): string
     {
-        return $this->id;
-    }
-
-    public function getPosition(): ?int
-    {
-        return $this->position;
-    }
-
-    public function setPosition(?int $position): static
-    {
-        $this->position = $position;
-
-        return $this;
-    }
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName(?string $name): static
-    {
-        $this->name = $name;
-
-        return $this;
-    }
-
-    public function getSize(): ?int
-    {
-        return $this->size;
-    }
-
-    public function setSize(?int $size): static
-    {
-        $this->size = $size;
-
-        return $this;
-    }
-
-    public function getFile(): ?string
-    {
-        return $this->file;
-    }
-
-    public function setFile(string $file): static
-    {
-        $this->file = $file;
-
-        return $this;
-    }
-
-    public function getUpdatedAt(): ?\DateTimeImmutable
-    {
-        return $this->updatedAt;
-    }
-
-    public function setUpdatedAt(\DateTimeImmutable $updatedAt): static
-    {
-        $this->updatedAt = $updatedAt;
-
-        return $this;
+        return 'crowdfundings';
     }
 
     public function getCrowdfunding(): ?Crowdfunding
     {
         return $this->crowdfunding;
     }
-
     public function setCrowdfunding(?Crowdfunding $crowdfunding): static
     {
         $this->crowdfunding = $crowdfunding;