+++ src/Entity/CrowdfundingVideo.php
@@ -1,96 +1,25 @@
 <?php
 
-/*
- * (c) 2025: 975L <contact@975l.com>
- * (c) 2025: Laurent Marquet <laurent.marquet@laposte.net>
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
 namespace c975L\ShopBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
-use c975L\ShopBundle\Repositiry\CrowdfundingVideoRepository;
+use c975L\ShopBundle\Entity\Crowdfunding;
 use Symfony\Component\HttpFoundation\File\File;
 use Vich\UploaderBundle\Mapping\Annotation as Vich;
 
-#[ORM\Entity(repositoryClass: CrowdfundingVideoRepository::class)]
-#[ORM\Table(name: 'shop_crowdfunding_video')]
+#[ORM\Entity]
 #[Vich\Uploadable]
-class CrowdfundingVideo
+class CrowdfundingVideo extends Media
 {
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column]
-    private ?int $id = null;
-
-    #[ORM\Column(length: 255, nullable: true)]
-    private ?string $name = null;
-
-    #[ORM\Column(nullable: true)]
-    private ?int $size = null;
-
     #[Vich\UploadableField(mapping: 'crowdfundings', fileNameProperty: 'name', size: 'size')]
-    private ?File $file = null;
+    protected ?File $file = null;
 
-    #[ORM\Column]
-    private ?\DateTimeImmutable $updatedAt = null;
-
-    #[ORM\OneToOne(mappedBy: 'video', cascade: ['persist', 'remove'])]
+    #[ORM\OneToOne(inversedBy: 'video', cascade: ['persist', 'remove'])]
     private ?Crowdfunding $crowdfunding = null;
 
-    public function getId(): ?int
+    public function getMappingName(): string
     {
-        return $this->id;
-    }
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName(?string $name): static
-    {
-        $this->name = $name;
-
-        return $this;
-    }
-
-    public function getSize(): ?int
-    {
-        return $this->size;
-    }
-
-    public function setSize(?int $size): static
-    {
-        $this->size = $size;
-
-        return $this;
-    }
-
-    public function getFile(): ?string
-    {
-        return $this->file;
-    }
-
-    public function setFile(string $file): static
-    {
-        $this->file = $file;
-
-        return $this;
-    }
-
-    public function getUpdatedAt(): ?\DateTimeImmutable
-    {
-        return $this->updatedAt;
-    }
-
-    public function setUpdatedAt(\DateTimeImmutable $updatedAt): static
-    {
-        $this->updatedAt = $updatedAt;
-
-        return $this;
+        return 'crowdfundings';
     }
 
     public function getCrowdfunding(): ?Crowdfunding