+++ src/Entity/ProductMedia.php
@@ -10,118 +10,26 @@
 
 namespace c975L\ShopBundle\Entity;
 
-use App\Entity\User;
 use Doctrine\ORM\Mapping as ORM;
+use c975L\ShopBundle\Entity\Media;
+use c975L\ShopBundle\Entity\Product;
 use Symfony\Component\HttpFoundation\File\File;
 use Vich\UploaderBundle\Mapping\Annotation as Vich;
-use c975L\ShopBundle\Repository\ProductMediaRepository;
 
-#[ORM\Entity(repositoryClass: ProductMediaRepository::class)]
-#[ORM\Table(name: 'shop_product_media')]
+#[ORM\Entity]
 #[Vich\Uploadable]
-class ProductMedia
+class ProductMedia extends Media
 {
-    #[ORM\Id]
-    #[ORM\GeneratedValue]
-    #[ORM\Column]
-    private ?int $id = null;
+    #[Vich\UploadableField(mapping: 'products', fileNameProperty: 'name', size: 'size')]
+    protected ?File $file = null;
 
-    #[ORM\Column(nullable: true)]
-    private ?int $position = null;
-
-    #[ORM\Column(nullable: true)]
-    private ?string $name = null;
-
-    #[ORM\Column(nullable: true)]
-    private ?int $size = null;
-
-    #[ORM\ManyToOne(inversedBy: 'medias', cascade: ['persist'])]
-    #[ORM\JoinColumn(nullable: false)]
+    #[ORM\ManyToOne(targetEntity: Product::class, inversedBy: 'medias')]
+    #[ORM\JoinColumn(nullable: true)]
     private ?Product $product = null;
 
-    #[Vich\UploadableField(mapping: 'products', fileNameProperty: 'name', size: 'size')]
-    private ?File $file = null;
-
-    #[ORM\Column]
-    private ?\DateTimeImmutable $updatedAt = null;
-
-    #[ORM\ManyToOne(inversedBy: 'productMedias')]
-    private ?User $user = null;
-
-    public function __toString()
+    public function getMappingName(): string
     {
-        return $this->name;
-    }
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getPosition(): ?int
-    {
-        return $this->position;
-    }
-
-    public function setPosition(?int $position): self
-    {
-        $this->position = $position;
-
-        return $this;
-    }
-
-    public function getName(): ?string
-    {
-        return $this->name;
-    }
-
-    public function setName(?string $name): static
-    {
-        $this->name = $name;
-
-        return $this;
-    }
-
-    public function getSize(): ?int
-    {
-        return $this->size;
-    }
-
-    public function setSize(?int $size): static
-    {
-        $this->size = $size;
-
-        return $this;
-    }
-
-    public function setFile(?File $file = null): static
-    {
-        $this->file = $file;
-
-        if (null !== $file) {
-            // It is required that at least one field changes if you are using doctrine
-            // otherwise the event listeners won't be called and the file is lost
-            $this->updatedAt = new \DateTimeImmutable();
-        }
-
-        return $this;
-    }
-
-    public function getFile(): ?File
-    {
-        return $this->file;
-    }
-
-    public function getUpdatedAt(): ?\DateTimeImmutable
-    {
-        return $this->updatedAt;
-    }
-
-    public function setUpdatedAt(\DateTimeImmutable $updatedAt): static
-    {
-        $this->updatedAt = $updatedAt;
-
-        return $this;
+        return 'products';
     }
 
     public function getProduct(): ?Product
@@ -132,18 +40,6 @@ class ProductMedia
     public function setProduct(?Product $product): static
     {
         $this->product = $product;
-
-        return $this;
-    }
-
-    public function getUser(): ?User
-    {
-        return $this->user;
-    }
-
-    public function setUser(?User $user): static
-    {
-        $this->user = $user;
 
         return $this;
     }