+++ src/Listener/ProductItemListener.php
@@ -10,27 +10,23 @@
 
 namespace c975L\ShopBundle\Listener;
 
-use DateTime;
+use DateTimeImmutable;
 use Doctrine\ORM\Events;
 use c975L\ShopBundle\Entity\ProductItem;
-use c975L\ShopBundle\Entity\ProductItemFile;
-use c975L\ShopBundle\Entity\ProductItemMedia;
+use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Event\PreFlushEventArgs;
 use Doctrine\ORM\Event\PrePersistEventArgs;
-use Doctrine\ORM\Event\PreRemoveEventArgs;
-use c975L\ShopBundle\Listener\Traits\MediaTrait;
-use c975L\ShopBundle\Listener\Traits\UserTrait;
-use Doctrine\Bundle\DoctrineBundle\Attribute\AsEntityListener;
-use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\SecurityBundle\Security;
+use c975L\ShopBundle\Entity\ProductItemFile;
+use c975L\ShopBundle\Entity\ProductItemMedia;
+use c975L\ShopBundle\Listener\Traits\UserTrait;
 use Symfony\Component\String\Slugger\SluggerInterface;
+use Doctrine\Bundle\DoctrineBundle\Attribute\AsEntityListener;
 
 #[AsEntityListener(event: Events::preFlush, method: 'preFlush', entity: ProductItem::class)]
 #[AsEntityListener(event: Events::prePersist, method: 'prePersist', entity: ProductItem::class)]
-#[AsEntityListener(event: Events::preRemove, method: 'preRemove', entity: ProductItem::class)]
 class ProductItemListener
 {
-    use MediaTrait;
     use UserTrait;
 
     public function __construct(
@@ -38,7 +34,6 @@ class ProductItemListener
         private readonly EntityManagerInterface $entityManager,
         private SluggerInterface $slugger,
     ) {
-        $this->initializeMedia($slugger);
     }
 
     public function preFlush(ProductItem $entity, PreFlushEventArgs $event): void
@@ -51,7 +46,7 @@ class ProductItemListener
             $entity->setPosition($maxPosition + 5);
         }
         $entity->setSlug($this->slugger->slug($entity->getTitle())->lower());
-        $entity->setModification(new DateTime());
+        $entity->setModification(new DateTimeImmutable());
         $this->setUser($entity);
     }
 
@@ -63,20 +58,16 @@ class ProductItemListener
         // Same for /ProductItemFile
         if (null === $entity->getMedia()) {
             $productItemMedia = new ProductItemMedia();
-            $productItemMedia->setUpdatedAt(new \DateTimeImmutable());
+            $productItemMedia->setUpdatedAt(new DateTimeImmutable());
+            $productItemMedia->setProductItem($entity);
             $entity->setMedia($productItemMedia);
         }
         if (null === $entity->getFile()) {
             $productItemFile = new ProductItemFile();
-            $productItemFile->setUpdatedAt(new \DateTimeImmutable());
+            $productItemFile->setUpdatedAt(new DateTimeImmutable());
+            $productItemFile->setProductItem($entity);
             $entity->setFile($productItemFile);
         }
-        $entity->setCreation(new DateTime());
-    }
-
-    public function preRemove(ProductItem $entity, PreRemoveEventArgs $event): void
-    {
-        // @TODO media file are not deleted... Can't find why (27/03/2025)
-        $this->deleteMedia($entity);
+        $entity->setCreation(new DateTimeImmutable());
     }
 }