+++ src/Listener/ProductListener.php
@@ -14,9 +14,7 @@ use DateTime;
 use Doctrine\ORM\Events;
 use c975L\ShopBundle\Entity\Product;
 use Doctrine\ORM\Event\PreFlushEventArgs;
-use Doctrine\ORM\Event\PreRemoveEventArgs;
 use Doctrine\ORM\Event\PrePersistEventArgs;
-use c975L\ShopBundle\Listener\Traits\MediaTrait;
 use c975L\ShopBundle\Listener\Traits\UserTrait;
 use Doctrine\Bundle\DoctrineBundle\Attribute\AsEntityListener;
 use Doctrine\ORM\EntityManagerInterface;
@@ -24,10 +22,8 @@ use Symfony\Bundle\SecurityBundle\Security;
 
 #[AsEntityListener(event: Events::preFlush, method: 'preFlush', entity: Product::class)]
 #[AsEntityListener(event: Events::prePersist, method: 'prePersist', entity: Product::class)]
-#[AsEntityListener(event: Events::preRemove, method: 'preRemove', entity: Product::class)]
 class ProductListener
 {
-    use MediaTrait;
     use UserTrait;
 
     public function __construct(
@@ -45,10 +41,5 @@ class ProductListener
     public function prePersist(Product $entity, PrePersistEventArgs $event): void
     {
         $entity->setCreation(new DateTime());
-    }
-
-    public function preRemove(Product $entity, PreRemoveEventArgs $event): void
-    {
-        $this->deleteMedias($entity);
     }
 }